api/requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
codecov==2.0.15
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
dj-database-url==0.5.0
Django==2.1.4
django-cors-headers==2.4.0
django-filter==2.0.0
django-redis==4.10.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
Faker==0.9.1
flake8==3.6.0
freezegun==0.3.11
gunicorn==19.9.0
idna==2.8
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
mixer==6.1.3
mock==2.0.0
more-itertools==4.3.0
pbr==5.1.1
Pillow==5.3.0
pluggy==0.8.0
psycopg2==2.7.6.1
py==1.7.0
pycodestyle==2.4.0
pyflakes==2.0.0
Pygments==2.3.0
PyJWT==1.7.1
pytest==4.0.1
pytest-cov==2.6.0
pytest-django==3.4.4
pytest-freezegun==0.3.0.post1
python-dateutil==2.7.5
pytz==2018.7
redis==3.0.1
requests==2.21.0
six==1.12.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
